src/examples/semantic-search-docs/requirements.txt
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
numpy>=1.20.0



